fix: isolate PostgREST features from supabase-js
Currently, some of postgrest-js's functionalities are implemented in
supabase-js, like
https://github.com/supabase/supabase-js/blob/33939998516e9325e207a63a8a3f10463e89aa41/src/index.js#L137
and
https://github.com/supabase/supabase-js/blob/33939998516e9325e207a63a8a3f10463e89aa41/src/index.js#L177.
This causes a mismatch of the implementations between the two, which
caused https://github.com/supabase/postgrest-js/issues/84.

Since supabase-js is made to act like postgrest-js, we should implement
PostgREST-related features solely on postgrest-js and expose that from
supabase-js through prototypal inheritance.