fix: rpc return type
RPC return types aren't getting propagated properly. E.g. if a function
returns `boolean`, the `data` will be typed as `boolean[]`.

To fix this, we change PostgrestBuilder's type parameter to be the
actual type of `data`, instead of the element type of `data` (which
assumes `data` is always an array, which is incorrect).

In doing so we effectively remove all usage of `PostgrestResponse` and
`PostgrestMaybeSingleResponse`. We can stop exporting these on the next
major version.

This also fixes the behavior of `.returns<NewResult>()`, which was
inconsistent with the documentation (i.e. `data` was typed as
`NewResult[]` instead of `NewResult`).