feat: add _acquireLock and navigatorLock (#736)
Built on top of: #734 

Adds an experimental `lock` option. This is used internally within
`_acquireLock()` which can be used to acquire a global exclusive lock
when using the client. It can prevent issues with concurrent access to
local storage.

`navigatorLock` is an experimental lock that can be used to test the
library before officially designating it (or an alternative) as the
default global exclusive locking mechanism. It will be rolled out on
supabase.com/dashboard over time as we grow our confidence in the
library.