realtime: Improve testing framework (#418)
We were using mocha which presented some challenges and less than good experience when it came to testing. We're moving to vitest and we're also using Typescript in our testing to easily catch some of the typing issues that might arise earlier.

We're also adding a coverage so later we can setup CI to take that into consideration.